requirements.txt
Django==1.10.5
django-choices==1.4.4
djangorestframework==3.5.3
six==1.10.0
